pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcpguard"
version = "0.1.0"
description = "Security scanner for MCP server configurations. Detects secrets, injection, auth issues, and maps to OWASP MCP Top 10."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "maxhu" }]
keywords = [
    "mcp", "security", "scanner", "claude", "ai",
    "vulnerability", "owasp", "model-context-protocol",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.urls]
Homepage = "https://github.com/loplop-h/mcpguard"
Issues = "https://github.com/loplop-h/mcpguard/issues"
Repository = "https://github.com/loplop-h/mcpguard"

[project.scripts]
mcpguard = "mcpguard.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mcpguard"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "UP", "SIM", "W", "RUF"]
